Remove init_dummy_user as it is not required
